+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Dijkstra/Ledger.hs
@@ -145,15 +145,16 @@ instance
   ) =>
   NFData (DijkstraStAnnTx SubTx era)
   where
-  rnf stAnnTx@(DijkstraStAnnSubTx _ _ _ _ _ _ _) =
+  rnf stAnnTx@(DijkstraStAnnSubTx _ _ _ _ _ _ _ _) =
     let DijkstraStAnnSubTx {..} = stAnnTx
      in dsastTx `deepseq`
           dsastScriptsNeeded `deepseq`
-            dsastScriptsProvided `deepseq`
-              dsastTxInfoResult `seq`
-                dsastPlutusLanguagesUsed `deepseq`
-                  dsastPlutusRunnableCache `deepseq`
-                    rnf dsastPlutusScriptsWithContext
+            dsastScriptsHashesNeeded `deepseq`
+              dsastScriptsProvided `deepseq`
+                dsastTxInfoResult `seq`
+                  dsastPlutusLanguagesUsed `deepseq`
+                    dsastPlutusRunnableCache `deepseq`
+                      rnf dsastPlutusScriptsWithContext
 
 instance EncCBOR (Tx TopTx era) => EncCBOR (DijkstraStAnnTx TopTx era) where
   encCBOR DijkstraStAnnTopTx {dsattTx} = encCBOR dsattTx
@@ -195,12 +196,13 @@ instance
   ) =>
   ToExpr (DijkstraStAnnTx SubTx era)
   where
-  toExpr stAnnTx@(DijkstraStAnnSubTx _ _ _ _ _ _ _) =
+  toExpr stAnnTx@(DijkstraStAnnSubTx _ _ _ _ _ _ _ _) =
     let DijkstraStAnnSubTx {..} = stAnnTx
      in TD.Rec "DijkstraStAnnSubTx" $
           OMap.fromList
             [ ("dsastTx", toExpr dsastTx)
             , ("dsastScriptsNeeded", toExpr dsastScriptsNeeded)
+            , ("dsastScriptsHashesNeeded", toExpr dsastScriptsHashesNeeded)
             , ("dsastScriptsProvided", toExpr dsastScriptsProvided)
             , ("dsastTxInfoResult", TD.App "<TxInfoResult>" [])
             , ("dsastPlutusRunnableCache", toExpr dsastPlutusRunnableCache)